Fix linker warnings of the kind: 1>LINK : warning LNK4075: ignoring '/INCREMENTAL' due to '/RELEASE' specification 1>acpienum.obj : warning LNK4075: ignoring '/EDITANDCONTINUE' due to '/DRIVER' specification
The /INCREMENTAL, /RELEASE warning occurs because the /SUBSYSTEM:NATIVE LINKER option is specified. The Linker knows that /SUBSYSTEM:NATIVE means that
the generated .SYS file must contain a valid checksum to be compliant with future revisions of the Windows Driver Model. To induce the programmer to 
be cognizant of compliance, /SUBYSTEM:NATIVE forces /RELEASE to be specified. /RELEASE means generate the checksum. Furthermore, to induce the 
programmer to think about code purity while build device driver, the Linker decides that incremental patching of the .SYS should not be allowed if
/RELEASE is specified, hence the warning.

The /EDITANDCONTINUE, /DRIVER error occurs for essentially the same reason. /EDITANDCONTINE, by design, relies on incremental linking. But as explained
above, if a device driver is being built, the linker wants the programmer to not have a patched .SYS, so /DRIVER disables incremental linking, which
disables /EDITANDCONTINUE.

The solution to fixing these two warnings is:

1. Set C/C++ -> General -> Debug Information Format : Program Database (/Zi)
2. Set Linker -> General -> Incremental Linking : No (/INCREMENTAL:NO)